mark all as read confirmation now also applies to the "shift+a" keyboard shortcut (#1744)
